% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{components}
\alias{components}
\title{Color palettes from components.ai, ramped to a specified length}
\usage{
components(palette = "lab", level = 8, n, alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Character. A palette to display; one of "bootstrap", "lab"
(the default), "material", "open_color", "palx", or "tachyons".}

\item{level}{Numeric. The "level" of the palette to be displayed.}

\item{n}{Numeric. Number of colors to be displayed.}

\item{alpha}{Numeric. Transparency.}

\item{reverse}{Logical. Should the order of colors be reversed?}
}
\description{
Color palettes from components.ai, ramped to a specified length
}
