% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{components_palettes}
\alias{components_palettes}
\title{Colors used in all Components Palettes}
\format{
A list of 6 elements, each a separate palette from components.ai.
Each palette is also included as a data frame on its own, and is documented
there.
}
\source{
\url{https://components.ai/theme}
}
\usage{
components_palettes
}
\description{
Colors used in all Components Palettes
}
\keyword{datasets}
