% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{components_show_palette}
\alias{components_show_palette}
\title{Display a color palette}
\usage{
components_show_palette(palette, level)
}
\arguments{
\item{palette}{vector of character hex RGB values}
}
\description{
Given a character vector (hex RGB values), display palette in graphics window.
}
