% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lab}
\alias{lab}
\title{Colors used in the Lab (Monotone) Palettes}
\format{
A data frame with 12 rows and 11 variables. Each row is a single
color palette, each column is a single color (with column names indicating
which color is represented), and each cell is a hexidecimal color code.
}
\source{
\url{https://components.ai/theme/sRr4Q9BGmH0jAOA99nc2}
}
\usage{
lab
}
\description{
Colors used in the Lab (Monotone) Palettes
}
\keyword{datasets}
