% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{material}
\alias{material}
\title{Colors used in the Material Palettes}
\format{
A data frame with 10 rows and 19 variables. Each row is a single
color palette, each column is a single color (with column names indicating
which color is represented), and each cell is a hexidecimal color code.
}
\source{
\url{https://components.ai/theme/material-2014}
}
\usage{
material
}
\description{
Colors used in the Material Palettes
}
\keyword{datasets}
