% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{open_color}
\alias{open_color}
\title{Colors used in the Open Color Palettes}
\format{
A data frame with 10 rows and 13 variables. Each row is a single
color palette, each column is a single color (with column names indicating
which color is represented), and each cell is a hexidecimal color code.
}
\source{
\url{https://components.ai/theme/open-color}
}
\usage{
open_color
}
\description{
Colors used in the Open Color Palettes
}
\keyword{datasets}
