% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palx}
\alias{palx}
\title{Colors used in the Palx Palettes}
\format{
A data frame with 10 rows and 13 variables. Each row is a single
color palette, each column is a single color (with column names indicating
which color is represented), and each cell is a hexidecimal color code.
}
\source{
\url{https://components.ai/theme/palx}
}
\usage{
palx
}
\description{
Colors used in the Palx Palettes
}
\keyword{datasets}
