% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{scale_color_components}
\alias{scale_color_components}
\alias{scale_colour_components}
\title{components.ai color scales for ggplot2}
\usage{
scale_color_components(
  palette = "lab",
  level = 8,
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)

scale_colour_components(
  palette = "lab",
  level = 8,
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character. A palette to display; one of "bootstrap", "lab"
(the default), "material", "open_color", "palx", or "tachyons".}

\item{level}{Numeric. The "level" of the palette to be displayed.}

\item{alpha}{Numeric. Transparency.}

\item{reverse}{Logical. Should the order of colors be reversed?}

\item{...}{Arguments passed to either [ggplot2::discrete_scale] or
[ggplot2::scale_color_gradientn], as appropriate.}
}
\description{
components.ai color scales for ggplot2
}
