% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tachyons}
\alias{tachyons}
\title{Colors used in the Tachyons Palettes}
\format{
A data frame with 10 rows and 16 variables. Each row is a single
color palette, each column is a single color (with column names indicating
which color is represented), and each cell is a hexidecimal color code.
}
\source{
\url{https://components.ai/theme/tachyons-v5}
}
\usage{
tachyons
}
\description{
Colors used in the Tachyons Palettes
}
\keyword{datasets}
